Create exportable tooltip component
- Replaces tooltip component in report.jsx
